requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
httpx>=0.27.0
pydantic>=2.0.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
langchain>=0.2.0
langchain-groq>=0.1.6
langchain-core>=0.2.0